October 1, 2014: 2:14pm
I found an error in the previous "Final Copy" of the HodgkinsHuxleyModel.m. This should be the absolute final copy of the model, and the one that should be graded.

I made a change to the Initialize Initial Injected Current section. The steady-state line was incorrect, and I fixed it.